Final rename and get docs structure in place
