+++ modules/server/src/index.js
@@ -11,5 +11,5 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-export {XVIZServer} from './server/server';
+export {XVIZServer} from './server/xviz-server';
 export {main} from './main';
