+++ test/modules/server/xviz-server.spec.js
